Move DHCP functionality in cmd/boots to dhcp/server package:
Move similar functionality/purpose together.
This will help with future refactors that will
ultimately change the underlining dhcp library.
